==> 03_Observations.Rmd <==
# Observation data

## Set-up

This example will use the following libraries:

```{r}
library(dismo)
library(sp)
library(here)
```

Load the shapefiles created earlier.
```{r}
load("data/shapefiles.RData")
```

## Download data

I will download occurrence data for *Trillium grandiflorum* and *Trillium undulatum* in my NHVT bounding box from the [Global Biodiversity Information Facility](https://www.gbif.org/). `nrecs` seems to be ignored. `geo` means only points with longitude and latitude. `removeZeros` means get rid of NA in location. `ext` is the bounding box to use. 

First I set where I will save the file and check if it is already there. I do this because if I rerun this script, I don't want to re-download. Note that GBIF data is updated weekly so using a time-stamp on your file might be good, but I am not doing that for this example.
```{r}
filePath <- file.path(here::here(), "data/trillium_presences.RData")
```
Now I download if I haven't downloaded already. The downloaded data has many columns that I don't need. I will subset the following columns. `select` in the `subset()` call says what columns to use. 
```{r}
if (!file.exists(filePath)) {
  # Download
  grandiflorum <- dismo::gbif("Trillium",
    species = "grandiflorum",
    nrecs = 300, geo = TRUE,
    removeZeros = TRUE, ext = NHVT
  )
  undulatum <- dismo::gbif("Trillium",
    species = "undulatum",
    nrecs = 300, geo = TRUE,
    removeZeros = TRUE, ext = NHVT
  )
  trillium.raw <- rbind(grandiflorum, undulatum)

  # select columns
  colsWeNeed <- c("species", "lat", "lon", "locality", "year", "coordinateUncertaintyInMeters", "occurrenceID", "occurrenceRemarks", "geodeticDatum")

  trillium.raw <- subset(trillium.raw, select = colsWeNeed)

  save(trillium.raw, file = "data/trillium_presences.RData")
}
```

Load in the presences data (saved from code above).
```{r}
load("data/trillium_presences.RData")
```

## Check the coordinate projection

Check the projection to make sure it makes sense and there is only one value. Check that it is the same projection as my other layers.
```{r}
unique(trillium.raw$geodeticDatum) # "WGS84"
```

## Make a sp object

`trillium.raw` is just a data frame. I make it a `sp` object (specifically a `SpatialPointsDataFrame`) using `sp::coordinates()` to specify which columns are the longitude and latitude.
```{r}
trillium <- trillium.raw
sp::coordinates(trillium) <- c("lon", "lat")
```
Check that it looks ok and there are no NAs.
```{r}
summary(trillium$lon)
summary(trillium$lat)
```


## Check for inaccurate location data

The `coordinateUncertaintyInMeters` column give the uncertainty of the observation location. Some of the uncertainties are huge and I don't want those.

```{r}
table(cut(trillium$coordinateUncertaintyInMeters, c(0, 200, 500, 1000, 2000, 5000)))
```
I am going to keep only those locations with a location accuracy within 200m.
```{r}
good <- which(trillium$coordinateUncertaintyInMeters < 200)
trillium <- trillium[good, ]
```

## Plot 

Now I can plot the occurrences points and add the NH and VT state boundaries. *Trillium undulatum* is much more common. Hubbard Brook is outlined in blue.
```{r}
plot(nhvtshp, border = "blue", axes = TRUE)
plot(subset(trillium, species == "Trillium grandiflorum"), pch = 3, cex = 1, add = TRUE)
plot(subset(trillium, species == "Trillium undulatum"), pch = 4, cex = 1, col = "red", add = TRUE)
plot(hbshp, add = TRUE, border = "blue")
```

## Save

```{r}
save(trillium, trillium.raw, file = "data/trillium_presences.RData")
```
